# swagger2.yml
consumes:
- application/json
definitions:
  Driver:
    description: '# Elementos de un driver'
    properties:
      dni:
        type: string
        x-go-name: Dni
      name:
        type: string
        x-go-name: Name
      phoneNumber:
        type: string
        x-go-name: PhoneNumber
    title: Driver
    type: object
    x-go-package: support-utils/structs/handlers
  DriverResponse:
    description: '# elemento de respuesta de drivers'
    properties:
      dni:
        type: string
        x-go-name: Dni
      name:
        type: string
        x-go-name: Name
      password:
        type: string
        x-go-name: Password
      phonenumber:
        type: string
        x-go-name: Phone
      username:
        type: string
        x-go-name: Username
    title: DriverResponse
    type: object
    x-go-package: support-utils/structs/responses
  DriverResponseArray:
    description: '# Array de drivers de respuesta'
    properties:
      drivers:
        items:
          $ref: '#/definitions/DriverResponse'
        type: array
        x-go-name: DResponse
    title: ADriverResponse
    type: object
    x-go-name: ADriverResponse
    x-go-package: support-utils/structs/responses
  WarehouseUsers:
    description: '# elemento de respuesta de warehouse users'
    properties:
      adm:
        format: int64
        type: integer
        x-go-name: Adm
      crd:
        format: int64
        type: integer
        x-go-name: Crd
      phone:
        type: string
        x-go-name: WarehousePhone
      pkr:
        format: int64
        type: integer
        x-go-name: Pkr
    title: WarehouseUsers
    type: object
    x-go-package: support-utils/structs/handlers
info:
  description: API Con las funcionalidades del equipo de soporte
  title: Support-utils.
paths:
  /acl/{warehouseCode}:
    put:
      description: '# Obtiene una frase aleatoria'
      operationId: DriverPost
      parameters:
      - in: body
        name: drivers
        schema:
          items:
            $ref: '#/definitions/Driver'
          type: array
        x-go-name: DriverA
      responses:
        "200":
          description: WarehouseUsers
          schema:
            items:
              $ref: '#/definitions/WarehouseUsers'
            type: array
  /driver/{warehouseCode}:
    post:
      description: '# Obtiene una frase aleatoria'
      operationId: DriverPost
      parameters:
      - in: body
        name: drivers
        schema:
          items:
            $ref: '#/definitions/Driver'
          type: array
        x-go-name: DriverA
      responses:
        "200":
          description: WarehouseUsers
          schema:
            items:
              $ref: '#/definitions/WarehouseUsers'
            type: array
produces:
- application/json
swagger: "2.0"